pyproject.toml
[project]
name = "notebooks"
version = "0.1.0"
description = "Collection of notebooks for learning new things"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncer>=0.0.9",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "cmd2>=2.5.3",
    "cryptography>=43.0.3",
    "duckdb>=1.4.0",
    "fastmcp>=2.11.3",
    "google-cloud-run>=0.11.0",
    "google-cloud-storage>=2.18.2",
    "h2tinker>=0.2",
    "inflect>=7.4.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "networkx>=3.4.2",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "playwright>=1.48.0",
    "prefect>=3.1.9",
    "prompt-toolkit>=3.0.48",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.9.2",
    "pydantic-ai>=0.4.3",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "pyx>=0.16",
    "qrcode>=8.2",
    "rich>=13.9.4",
    "scapy>=2.6.1",
    "sqlalchemy>=2.0.35",
    "strawberry-graphql[debug-server]>=0.270.3",
    "twilio>=9.8.4",
    "ulid>=1.1",
    "uuid-utils>=0.11.1",
]
